Error bar support for the histogram plot.
In the histogram syle configuration there will be an option to overlay
error bars on top of the existing plot.  This uses the flot errorbars
plugin. The error bars are disbled when derivatives are enabled.

The errors are computed from the min and max values returned by ES and
comparing it against whatever chart value is currently active. Not sure
that it makes much sense to plot errors against anything other than mean.